drop python 3.5 support
* adjust CIs & wheel scripts
* remove python & C++ code for python 3.5
