+++ CMakeLists.txt
@@ -332,7 +332,7 @@ endif()
 # =============================================================================
 # Enable Python support
 # =============================================================================
-find_package(PythonInterp REQUIRED)
+find_package(PythonInterp 3.6 REQUIRED)
 if(NRN_ENABLE_PYTHON)
   # start afresh with PythonLibsNew's find_library
   # without this, -DPYTHON_EXECUTABLE=`which python3` -DNRN_PYTHON_DYNAMIC="python3.6;python3.9"
@@ -680,16 +680,13 @@ if(NRN_ENABLE_PYTHON)
   message(STATUS "  LIB         | ${NRN_DEFAULT_PYTHON_LIBRARIES}")
   message(STATUS "  MODULE      | ${NRN_ENABLE_MODULE_INSTALL}")
   message(STATUS "  DYNAMIC     | ${NRN_ENABLE_PYTHON_DYNAMIC}")
-  if(PYTHON_VERSION_MAJOR LESS 3)
-    message(SEND_ERROR "Python 3 required. Please upgrade.")
-  endif()
   if(NRN_ENABLE_PYTHON_DYNAMIC)
     list(LENGTH NRN_PYTHON_EXE_LIST _num_pythons)
     math(EXPR num_pythons "${_num_pythons} - 1")
     foreach(val RANGE ${num_pythons})
       list(GET NRN_PYTHON_EXE_LIST ${val} exe)
       list(GET NRN_PYTHON_VER_LIST ${val} version)
-      if(PYTHON_VERSION_MAJOR LESS 3)
+      if(${version} LESS 3)
         message(SEND_ERROR "Python 3 required. Please upgrade.")
       endif()
       list(GET NRN_PYTHON_INCLUDE_LIST ${val} include)
