+++ packaging/python/build_wheels.bash
@@ -9,7 +9,7 @@ set -xe
 #  - cmake (>=3.5)
 #  - flex
 #  - bison
-#  - python >= 3.5
+#  - python >= 3.6
 #  - cython
 #  - MPI
 #  - X11
@@ -49,7 +49,6 @@ pip_numpy_install() {
     # numpy is special as we want the minimum wheel version
     numpy_ver="numpy"
     case "$py_ver" in
-      35) numpy_ver="numpy==1.10.4" ;;
       36) numpy_ver="numpy==1.12.1" ;;
       37) numpy_ver="numpy==1.14.6" ;;
       38) numpy_ver="numpy==1.17.5" ;;
@@ -68,7 +67,7 @@ build_wheel_linux() {
 
     echo " - Installing build requirements"
     #auditwheel needs to be installed with python3
-    PATH=/opt/python/cp38-cp38/bin/:$PATH pip3 install auditwheel
+    pip install auditwheel
     pip install -r packaging/python/build_requirements.txt
     pip_numpy_install
 
@@ -88,7 +87,7 @@ build_wheel_linux() {
         echo " - Auditwheel show"
         PATH=/opt/python/cp38-cp38/bin/:$PATH auditwheel show dist/*.whl
         echo " - Repairing..."
-        PATH=/opt/python/cp38-cp38/bin/:$PATH auditwheel repair dist/*.whl
+        auditwheel repair dist/*.whl
     fi
 
     deactivate
