+++ packaging/python/test_wheels.sh
@@ -166,12 +166,6 @@ else
   echo " == Using global install == "
 fi
 
-# on osx we need to install pip from source
-if [[ "$OSTYPE" == "darwin"* ]] && [[ "$python_ver" == "35" ]]; then
-  echo "Updating pip for OSX with Python 3.5"
-  curl https://raw.githubusercontent.com/pypa/get-pip/20.3.4/get-pip.py | python
-fi
-
 # install neuron and neuron
 $python_exe -m pip install numpy
 $python_exe -m pip install $python_wheel
